Extract shared architecture helpers, add PNA e2e tests
PSA and PNA duplicated ~350 lines of architecture-agnostic code. Extract
into shared files so adding future architectures is even cheaper:
- ArchitectureHelpers.kt: pipeline infrastructure (stage runners, param
  binding, default values, action selector forks)
- TraceHelpers.kt: buildForkTree
- Hash.kt: evalGetHash (parameterized on algorithm map), hashDataArg, sumWords

Also add 3 PNA corpus STF tests exercising send_to_port, drop_packet,
drop-by-default, LPM routing, and direct counters.
